refactor(react-native): remove the need to monkey patch the attach platform extension
