+++ packages/platforms/react-native/lib/platform-extensions.tsx
@@ -1,20 +1,22 @@
-import type { Clock, SpanContextStorage, SpanOptions } from '@bugsnag/core-performance'
+import type { Client, Clock, SpanContextStorage, SpanOptions } from '@bugsnag/core-performance'
 import React from 'react'
-import type { ReactNativeAttachConfiguration } from './config'
+import { Platform } from 'react-native'
+import NativeBugsnagPerformance from './native'
+import type { ReactNativeAttachConfiguration, ReactNativeConfiguration, ReactNativeSchema } from './config'
 import { createAppStartSpan } from './create-app-start-span'
 import { createNavigationSpan } from './create-navigation-span'
 import type { ReactNativeSpanFactory } from './span-factory'
 
 type NavigationSpanOptions = Omit<SpanOptions, 'isFirstClass'>
 
-export const platformExtensions = (appStartTime: number, clock: Clock, spanFactory: ReactNativeSpanFactory, spanContextStorage: SpanContextStorage) => ({
-  startNavigationSpan: (routeName: string, spanOptions?: NavigationSpanOptions) => {
+export const platformExtensions = (appStartTime: number, clock: Clock, schema: ReactNativeSchema, spanFactory: ReactNativeSpanFactory, spanContextStorage: SpanContextStorage) => ({
+  startNavigationSpan: function (routeName: string, spanOptions?: NavigationSpanOptions) {
     const cleanOptions = spanFactory.validateSpanOptions(routeName, spanOptions)
     cleanOptions.options.isFirstClass = true
     const span = createNavigationSpan(spanFactory, cleanOptions.name, cleanOptions.options)
     return spanFactory.toPublicApi(span)
   },
-  withInstrumentedAppStarts: (App: React.FC) => {
+  withInstrumentedAppStarts: function (App: React.FC) {
     const appStartSpan = createAppStartSpan(spanFactory, appStartTime)
 
     return () => {
@@ -27,8 +29,27 @@ export const platformExtensions = (appStartTime: number, clock: Clock, spanFacto
       return <App />
     }
   },
-  attach: (config?: ReactNativeAttachConfiguration) => {
-    // this noop implementation is overridden by the client and is just defined here for correct typing
+  attach: function (config?: ReactNativeAttachConfiguration) {
+    const logger = schema.logger.validate(config?.logger) ? config.logger : schema.logger.defaultValue
+    const isNativePerformanceAvailable = NativeBugsnagPerformance !== null && NativeBugsnagPerformance.isNativePerformanceAvailable()
+    if (!isNativePerformanceAvailable) {
+      logger.warn(`Could not attach to native SDK. No compatible version of Bugsnag ${Platform.OS} performance was found`)
+      return
+    }
+
+    const nativeConfig = NativeBugsnagPerformance?.getNativeConfiguration()
+    if (!nativeConfig) {
+      logger.warn(`Could not attach to native SDK. Bugsnag ${Platform.OS} performance has not been started`)
+      return
+    }
+
+    const finalConfig: ReactNativeConfiguration = {
+      ...config,
+      ...nativeConfig
+    }
+
+    const client = this as unknown as Client<ReactNativeConfiguration>
+    client.start(finalConfig)
   }
 })
